+++ CHANGES.md
@@ -7,6 +7,8 @@
 - Add a code action to open the closest Dune file for the current document.
   (#1817, fixes #1491, @rgrinberg)
 - Distinguish operators from functions in semantic highlighting. (#1831, @rgrinberg)
+- Classify function parameters and labels in semantic highlighting.
+  (#1857, @rgrinberg)
 
 ## Fixes